JWT utilities
Migrated the authentication method from tower_sessions to the
authentication services offered by Supabase. This involves an access and
refresh JWT stored by the user. Varying levels of authentication will
exist to check some of two conditions
1. (weak) the access JWT is valid and has not been tampered with
2. (strong) the access JWT refers to a session that has not been
   terminated due either to expiration or logout
It is expected that both access and refresh JWTs are sent back to the
client, and that the client handles refresh whenever it identifies that
the access JWT has expired. The access JWT is expected to be found in
the Authorization header, in form:
  Authorization: Bearer <token>

Utilities revolving around handling JWTs are introduced in this commit.