auth architecture
Redesigned the auth/ crate so that it is self contained. The design
necessitates only a call to receive the authentication router and
middleware -- all other implementation details are, or will be changed
to become, hidden.

Introduced the AuthManager trait representing structures capable of
handling key authentication functionality. This will allow for
extensibility in the situation where the authentication provider is
changed or recreated.