auth complete
The auth package is complete, and is ready to be merged into the main
branch. This commit simply removed several TODO comments requesting for
tests -- the tests would be delegated to the Authenticator, however the
supabase authenticator used by default defers all logic to the
supabase_auth crate, so that functionality is tested in a separate
crate.